added delete emploeee && added delete reciept

+++ lib/src/core/services/db/remote.db.services.dart
@@ -78,12 +78,16 @@ class MongoDataBase {
     return _products.deleteMany({
       '_id': {'\$in': model.productIds?.map((e) => ObjectId.fromHexString(e ?? '')).toList()}
     }).then((e) {
-      if (e.isSuccess) {
-        return _categories.deleteMany({'_id': ObjectId.fromHexString(model.id ?? '')}).then((e) => e.isSuccess);
-      }
+      if (e.isSuccess) return _categories.deleteMany({'_id': ObjectId.fromHexString(model.id ?? '')}).then((e) => e.isSuccess);
       return false;
     });
   }
-  // Future<bool> deleteOneEmployee(id) => _employees.findOne({'_id': id}).then((e) => true);
-  // Future<bool> deleteOneReciept(id) => _reciepts.findOne({'_id': id}).then((e) => true);
+
+  Future<bool> deleteOneEmployee(EmployeeModel model) {
+    return _employees.deleteOne({'_id': ObjectId.fromHexString(model.id ?? '')}).then((e) => e.isSuccess);
+  }
+
+  Future<bool> deleteOneReciept(RecieptModel model) {
+    return _reciepts.deleteOne({'_id': ObjectId.fromHexString(model.id ?? '')}).then((e) => e.isSuccess);
+  }
 }
